+++ dev/modules/Threading/qtp_worker.py
@@ -51,7 +51,7 @@ class Worker(QRunnable):
 
         # Add the callback to our kwargs
         self.kwargs['progress_callback'] = self.signals.progress
-        # Add the get_data function to our kwargs
+        # Add the get_data function to enable requesting input
         self.kwargs['get_data'] = self.get_input
         self.signals.data_response.connect(self.set_input)
         self.user_input = None
