+++ src/main/gui_main.py
@@ -14,7 +14,8 @@ from pathlib import Path
 from pydantic import ValidationError
 
 from assethandling.asset_manager import settings
-from assethandling.basemodels import UtilTabInput, RawTabInput, FolderTabInput, ExcelOptions, ExcelInput
+from assethandling.basemodels import UtilTabInput, RawTabInput, FolderTabInput,\
+    ExcelOptions, ExcelInput
 from ui import Ui_MainWindow, messageboxes, SelectionDialog
 from src.main.threadworker.thread_worker import Worker
 from src.main.runner import runners
@@ -158,6 +159,8 @@ class MainWindow(QMainWindow):
         # ##### BUTTONS "Ordner erstellt" ##### #
         self.ui.pb_foldersetup.clicked.connect(
             lambda: self.run_folder_setup(function=runners.run_folder_setup))
+
+
         # ##### BUTTONS "Rohmaterial verarbeiten" ##### #
         # todo
         self.ui.pb_start_raw_full.clicked.connect(
@@ -175,6 +178,8 @@ class MainWindow(QMainWindow):
         self.ui.pb_create_picture_folder.clicked.connect(
             lambda: self.run_raw_action(function=runners.run_create_picture_folder)
         )
+
+
         # ##### BUTTONS "Auswertung" ##### #
         self.ui.pb_section.clicked.connect(
             lambda: self.run_util_action(function=runners.run_copy_sections))
@@ -306,8 +311,11 @@ class MainWindow(QMainWindow):
             return
         self.run_action(function=function, slot=self.write_process_setup, input_=data)
 
-    # todo
-    def get_raw_input(self, excel=None) -> RawTabInput:
+    # ------------------------- todo ------------------
+    def get_raw_input_new(self) -> RawTabInput:
+        # get all ui informations
+        # deal with logics later
+        # how to handle excel
         data = {
             "do_structure": self.ui.cb_structure.isChecked(),
             "do_rename": self.ui.cb_rename.isChecked(),
@@ -315,84 +323,20 @@ class MainWindow(QMainWindow):
             "create_picture_folder": self.ui.cb_diashow.isChecked(),
             "raw_material_folder": Path(self.ui.drop_raw_rawpath.text()),
             "first_folder_date": self.ui.date_correct_fs.date().toPyDate(),
-            "excel": excel,
+            # "excel": excel,
         }
-        if self.ui.custom_picture_folder.isChecked():
-            if self.ui.drop_picture_folder.text() == "":
-                raise ValueError("Bitte gib einen Speicherort für den Bilderordner an.")
-            data["picture_folder"] = Path(self.ui.drop_picture_folder.text())
-        else:
-            data["picture_folder"] = data["raw_material_folder"].parent / "Bilderordner"
-
         return RawTabInput(**data)
 
-    def get_excel_data(self, override) -> Path | ExcelInput | None:
-        if not self.ui.cb_fill_excel:
-            return None
-        excel_option: ExcelOptions = ExcelOptions(self.ui.excel_option.currentText())
-        if excel_option == ExcelOptions.EXISTING:
-            return Path(self.ui.drop_raw_excel_file.text())
-        else:
-            try:
-                config = self._get_excel_input(option=excel_option)
-                config.override = override
-                if (config.excel_folder / config.excel_file_name).exists() and not override:
-                    self.open_excel_exists()
-                    return
-                return config
-            except Exception as e:
-                traceback.print_exc()
-                self.open_problem_input(str(e))
-
-    def _get_excel_input(self, option) -> ExcelInput:
-        if option != ExcelOptions.STANDARD and option != ExcelOptions.MANUAL:
-            error = "Interner Fehler, der Button 'Excel erstellen'\n" \
-                    "sollte nicht klickbar sein mit der Option \n" \
-                    "vorhandene Excel nutzen. Neustarten."
-            raise ValueError(error)
-        elif self.ui.drop_raw_rawpath.text() == "":
-            error = "Bitte gib einen Speicherort für die Excel-Datei an.\n" \
-                    "Dazu kannst du bei 'Excel-Datei' einen Ordner \n" \
-                    "angeben oder für den Standardweg den Rohmaterialorder \nangeben." \
-                    "Für Standards schau dir doch gerne die Anleitung an."
-            raise ValueError(error)
-        else:
-            if option == ExcelOptions.MANUAL:
-                config = ExcelInput(
-                    excel_folder=Path(self.ui.drop_excel_folder.text()),
-                    excel_file_name=self.ui.le_excel_file_name.text(),
-                    video_columns=self.get_PlainTextEdit_parts(self.ui.vid_columns),
-                    picture_columns=self.get_PlainTextEdit_parts(self.ui.pic_columns)
-                )
-            else:
-                config = ExcelInput(
-                    excel_folder=Path(self.ui.drop_raw_rawpath.text())
-                )
-            return config
-
-    # todo
-    def run_raw_action(self, function):
+    def run_raw_action_new(self, function):
         try:
-            self.current_function = function
             data: RawTabInput = self.get_raw_input()
         except (ValidationError, ValueError) as e:
             traceback.print_exc()
             self.open_problem_input(error=str(e))
             return
-        self.current_function = None
         self.run_action(function=function, slot=self.write_process_raw, input_=data)
 
-    # todo
-    def raw_with_excel(self, function, override=False):
-        try:
-            self.current_function = function
-            data: RawTabInput = self.get_raw_input(self.get_excel_data(override=override))
-        except (ValidationError, ValueError) as e:
-            traceback.print_exc()
-            self.open_problem_input(error=str(e))
-            return
-        self.current_function = None
-        self.run_action(function=function, slot=self.write_process_raw, input_=data)
+    # ------------------------- to do ------------------
 
     def run_util_action(self, function):
         try:
@@ -428,7 +372,6 @@ class MainWindow(QMainWindow):
         }
         return UtilTabInput(**data)
 
-    # todo add get_data
     def run_action(self, function, slot, input_):
         self.disable_work_buttons()
         self.movie.start()
@@ -495,11 +438,13 @@ class MainWindow(QMainWindow):
                 cols.append(part.strip())
         return cols
 
+
 def main():
     app = QApplication(sys.argv)
     window = MainWindow()
     window.show()
     sys.exit(app.exec())
+
 
 if __name__ == '__main__':
     main()
